requirements.txt
flask
flask-admin
pymysql
flask-sqlalchemy
flask-script
